Trains/Tests/RefereeTests.java
import Admin.Referee;
import Common.PlayerGameState;
import Common.TrainsMap;
import Other.*;
import Player.Player;
import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.ExpectedException;

import java.util.*;

import static junit.framework.TestCase.*;

/**
 * Tests for a Referee
 */
public class RefereeTests {
    @Rule
    public final ExpectedException expectedEx = ExpectedException.none();

    TestUtils utils = new TestUtils();

    private TrainsMap map;
    private List<Integer> orderOfPlayerTurns;
    private int currentPlayer;
    Map<Connection, Integer> connectionsStatus;

    private Set<Destination> destinationsToPickFrom = new HashSet<>();
    Map<ConnectionColor, Integer> playerCards;

    PlayerInventory inventory1;
    PlayerInventory inventory2;

    PlayerGameState playerGameState1;
    PlayerGameState playerGameState2;

    Player player1;
    Player player2;
    Player player3;
    Player player4;

    List<Player> players1 = new ArrayList<>();
    List<Player> players2 = new ArrayList<>();

    Referee referee;
    Referee referee2;


    @Before
    public void setUpReferee() {
        setupPlayers();
        players1.add(player1);
        players1.add(player2);

        players2.add(player1);
        players2.add(player3);
        players2.add(player2);
        players2.add(player4);

        referee = new Referee(map, players1);
        referee2 = new Referee(map, players2);
    }

    /**
     * Sets up two players.
     */
    private void setupPlayers() {
        map = utils.map;
        orderOfPlayerTurns = utils.playerTurns;
        currentPlayer = orderOfPlayerTurns.get(0);
        connectionsStatus = utils.createConnectionsStatus(-1);
        playerCards = new HashMap<>();

        playerCards.put(ConnectionColor.red, 1);
        playerCards.put(ConnectionColor.green, 1);
        playerCards.put(ConnectionColor.white, 1);
        playerCards.put(ConnectionColor.blue, 6);

        this.inventory1 = new PlayerInventory(new HashMap<>(playerCards), 40, new HashSet<>());
        this.inventory2 = new PlayerInventory(new HashMap<>(playerCards), 40, new HashSet<>());

        playerGameState1 = new PlayerGameState(map, orderOfPlayerTurns, currentPlayer, connectionsStatus, inventory1);
        playerGameState2 = new PlayerGameState(map, orderOfPlayerTurns, currentPlayer, connectionsStatus, inventory2);

        player1 = new Player(15, playerGameState1,"$PWD/Trains/Other/Strategies/BuyNowStrategy.java");
        player2 = new Player(56, playerGameState1,"$PWD/Trains/Other/Strategies/Hold10Strategy.java");
        player3 = new Player(-10, playerGameState1,"$PWD/Trains/Other/Strategies/MockDestinationCheatingStrategy.java");
        player4 = new Player(0, playerGameState1,"$PWD/Trains/Other/Strategies/MockAcquiringCheatingStrategy.java");

        destinationsToPickFrom = utils.createNewDestinationsToPickFrom();
    }

    /**
     * Creates a new referee with the given players
     * @param players a list of Player
     * @return a new Referee
     */
    private Referee createReferee(List<Player> players) {
        return new Referee(map, players);
    }

    @Test
    public void testCheatingDestinationPlayerEliminated() {
        List<Player> expectedEliminated = new ArrayList<>();
        expectedEliminated.add(player3);
        expectedEliminated.add(player4);

        referee2.playGame();
        assertEquals(expectedEliminated, referee2.getEliminatedPlayers());
    }

    @Test
    public void testNonCheatingPlayersNotEliminated() {
        List<Player> expectedNotEliminated = new ArrayList<>();
        expectedNotEliminated.add(player1);
        expectedNotEliminated.add(player2);
        referee2.playGame();
        assertEquals(expectedNotEliminated, referee2.getActivePlayers());
    }

    @Test
    public void testRefereeDuplicatePlayers() {
        expectedEx.expect(IllegalArgumentException.class);
        expectedEx.expectMessage("List of players contains duplicate players");

        List<Player> duplicatePlayers = new ArrayList<>();
        duplicatePlayers.add(player1);
        duplicatePlayers.add(player1);
        createReferee(duplicatePlayers);
    }

    @Test
    public void testRefereeWrongNumPlayers() {
        expectedEx.expect(IllegalArgumentException.class);
        expectedEx.expectMessage("Number of players must be between 2 and 8");

        List<Player> onePlayer = new ArrayList<>();
        onePlayer.add(player1);
        createReferee(onePlayer);
    }

    @Test
    public void testPlayGameUpdatesGameState() {
        assertEquals(referee.getGameState().getInventories(), new HashMap<>());
        referee.playGame();
        assertEquals(player1.getInventory().getDestinations(),
                referee.getGameState().getInventories().get(player1.getPlayerID()).getDestinations());
    }

    @Test
    public void testPlayerGameUPdatesGameState2() {
        // TODO: debug for next milestone. Player inventories inside a player game
        //  state should look like the inventories inside a referee game state.
        referee.playGame();
        /*
        assertEquals(player1.getInventory().getCardsInHand(),
                referee.getGameState().getInventories().get(player1.getPlayerID()).getCardsInHand());
        assertEquals(player1.getInventory().getRailsInHand(),
                referee.getGameState().getInventories().get(player1.getPlayerID()).getRailsInHand());
        assertEquals(player1.getInventory().getDestinations(),
                referee.getGameState().getInventories().get(player1.getPlayerID()).getDestinations());
        assertEquals(player2.getInventory().getCardsInHand(),
                referee.getGameState().getInventories().get(player2.getPlayerID()).getCardsInHand());
        assertEquals(player2.getInventory().getRailsInHand(),
                referee.getGameState().getInventories().get(player2.getPlayerID()).getRailsInHand());
        assertEquals(player2.getInventory().getDestinations(),
                referee.getGameState().getInventories().get(player2.getPlayerID()).getDestinations());
         */
    }

    @Test
    public void testGoodGame() {
        List<Player> players = new ArrayList<>();
        players.add(player1);
        players.add(player2);
        Referee ref = createReferee(players);
        assertEquals(players, ref.getActivePlayers());
        assertEquals(new ArrayList<>(), ref.getEliminatedPlayers());
    }

    @Test
    public void testRankingsGameNotPlayedYet() {
        List<Player> players = new ArrayList<>();
        players.add(player1);
        players.add(player2);
        Referee ref = createReferee(players);

        List<Set<Player>> ranking = ref.getRanking();
        Set<Player> playerSet = new HashSet<>();
        playerSet.add(player1);
        playerSet.add(player2);

        assertEquals(ranking.get(0), playerSet);
    }

    @Test
    public void testRankings() {
        List<Player> players = new ArrayList<>();
        players.add(player1);
        players.add(player2);
        Referee ref = createReferee(players);

        ref.playGame();

        List<Set<Player>> ranking = ref.getRanking();
        Set<Player> player1Set = new HashSet<>();
        player1Set.add(player1);
        Set<Player> player2Set = new HashSet<>();
        player2Set.add(player2);
        Set<Player> player12Set = new HashSet<>();
        player12Set.add(player1);
        player12Set.add(player2);

        if (ranking.size() == 2) {
            assertTrue(ranking.get(0).equals(player1Set) || ranking.get(0).equals(player2Set));
        } else if (ranking.size() == 1){
            assertTrue(ranking.get(0).equals(player12Set));
        }
    }

    @Test
    public void testRankingsEliminatedPlayers() {
        List<Player> players = new ArrayList<>();
        players.add(player1);
        players.add(player3);
        Referee ref = createReferee(players);

        ref.playGame();

        List<Set<Player>> ranking = ref.getRanking();
        Set<Player> player1Set = new HashSet<>();
        player1Set.add(player1);
        Set<Player> player3Set = new HashSet<>();
        player3Set.add(player3);

        assertEquals(ranking.get(0), player1Set);
        assertEquals(ranking.get(1), player3Set);
    }

    @Test
    public void testDFSAcc() {
        City a = new City("A", 0.1, 0.2);
        City b = new City("B", 0.12, 0.23);
        City c = new City("C", 0.21, 0.34);
        City d = new City("D", 0.81, 0.21);
        City e = new City("E", 0.9, 0.31);
        City f = new City("F", 0.91, 0.3);

        Set<City> cities = new HashSet<>();
        cities.add(a);
        cities.add(b);
        cities.add(c);
        cities.add(d);
        cities.add(e);
        cities.add(f);

        Set<Connection> connections = new HashSet<>();
        connections.add(new Connection(ConnectionColor.blue, 3, a, b));
        connections.add(new Connection(ConnectionColor.red, 4, c, d));
        connections.add(new Connection(ConnectionColor.white, 3, d, a));
        connections.add(new Connection(ConnectionColor.blue, 5, d, e));
        connections.add(new Connection(ConnectionColor.green, 5, e, c));
        connections.add(new Connection(ConnectionColor.green, 5, e, b));
        connections.add(new Connection(ConnectionColor.green, 4, b, f));
        //assertEquals(20, referee.DFS(cities, connections)); // ad, dc ce eb bf = 3 4 5 5 4 = 21

        Set<Connection> connections2 = new HashSet<>();
        connections2.add(new Connection(ConnectionColor.blue, 3, a, b));
        connections2.add(new Connection(ConnectionColor.red, 4, c, d));
        connections2.add(new Connection(ConnectionColor.white, 3, d, a));
        connections2.add(new Connection(ConnectionColor.blue, 5, d, e));
        connections2.add(new Connection(ConnectionColor.green, 5, e, c));
        connections2.add(new Connection(ConnectionColor.green, 4, e, b));
        //assertEquals(16, referee.DFS(cities, connections2)); // ad, dc ce eb = 16
    }
}
